+++ src/main/java/com/brokerage/service/OrderService.java
@@ -6,5 +6,5 @@ import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface OrderService {
-    Order createOrder(UUID customerId, String assetName, String orderSide, int size, BigDecimal price);
+    Order createOrder(UUID customerId, String assetName, String orderSide, BigDecimal size, BigDecimal price);
     }
